Labfiles/06-build-remote-agents-with-a2a-es/python/requirements.txt
python-dotenv
azure-identity
a2a-sdk
httpx
requests
fastapi
uvicorn
starlette
